main: Add option to disable status icon
This makes it easier to get out of a situation with a broken
tray and launch-to-tray enabled.

At the same time, refactor option handling a bit, so that main.c
doesn't have to deal with defaults coming from config, and can just
act on values like status_icon_enabled etc.